Se configura cliente retrofit y se agrega logica de consumo de api de peliculas

+++ domain/src/main/java/com/jeferson/chiper/android/moviedb/domain/MovieDomain.kt
@@ -13,6 +13,6 @@ data class MovieDomain(
     val release_date: String?,
     val title: String,
     val video: Boolean,
-    val vote_average: Float?,
+    val vote_average: Double?,
     val vote_count: Int,
 )
